optimize parse layer
This commit change the internal storage of the signatures in InputEvent.
We merge all signatures into a single string as the old ipmes did. Along
with using one regex matcher to match all signatures, this increase the
matching performance.

In addition, I manually write the input graph parsing code to imporve
performance further.